Integrate with SeaORM DB connections and a Database Config
