+++ examples/async-graphql/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nakago-examples-async-graphql"
-version = "0.2.1"
+version = "0.8.0"
 authors = ["Brandon Konkle <brandon@konkle.us>"]
 edition = "2021"
 description = "A lightweight Rust framework for sharp services 😎"
@@ -24,6 +24,7 @@ log = "0.4"
 once_cell = "1.9"
 nakago = "0.7"
 nakago-axum = "0.7"
+nakago-sea-orm = "0.8"
 oso = "0.27"
 pico-args = "0.5.0"
 pretty_env_logger = "0.5"
