+++ examples/async-graphql/src/domains/episodes/service.rs
@@ -6,7 +6,8 @@ use async_trait::async_trait;
 #[cfg(test)]
 use mockall::automock;
 use nakago::{Dependency, Inject, InjectResult, Provider, Tag};
-use sea_orm::{entity::*, query::*, DatabaseConnection, EntityTrait};
+use nakago_sea_orm::{DatabaseConnection, DATABASE_CONNECTION};
+use sea_orm::{entity::*, query::*, EntityTrait};
 
 use super::{
     model::{self, Episode, EpisodeList, EpisodeOption},
@@ -14,7 +15,6 @@ use super::{
     queries::{EpisodeCondition, EpisodesOrderBy},
 };
 use crate::{
-    db::DATABASE_CONNECTION,
     domains::shows::model as show_model,
     utils::{ordering::Ordering, pagination::ManyResponse},
 };
