+++ examples/async-graphql/src/domains/profiles/service.rs
@@ -6,7 +6,8 @@ use async_trait::async_trait;
 #[cfg(test)]
 use mockall::automock;
 use nakago::{Dependency, Inject, InjectResult, Provider, Tag};
-use sea_orm::{entity::*, query::*, DatabaseConnection, EntityTrait};
+use nakago_sea_orm::{DatabaseConnection, DATABASE_CONNECTION};
+use sea_orm::{entity::*, query::*, EntityTrait};
 
 use super::{
     model::{self, Profile, ProfileList, ProfileOption},
@@ -14,7 +15,6 @@ use super::{
     queries::{ProfileCondition, ProfilesOrderBy},
 };
 use crate::{
-    db::DATABASE_CONNECTION,
     domains::users::model as user_model,
     utils::{ordering::Ordering, pagination::ManyResponse},
 };
