+++ examples/async-graphql/src/domains/role_grants/service.rs
@@ -5,9 +5,8 @@ use async_trait::async_trait;
 #[cfg(test)]
 use mockall::automock;
 use nakago::{Dependency, Inject, InjectResult, Provider, Tag};
-use sea_orm::{entity::*, query::*, Condition, DatabaseConnection, EntityTrait};
-
-use crate::db::DATABASE_CONNECTION;
+use nakago_sea_orm::{DatabaseConnection, DATABASE_CONNECTION};
+use sea_orm::{entity::*, query::*, Condition, EntityTrait};
 
 use super::model::{self, CreateRoleGrantInput, RoleGrant};
