+++ examples/async-graphql/src/utils/mod.rs
@@ -7,8 +7,5 @@ pub mod ordering;
 /// Pagination Utils
 pub mod pagination;
 
-/// Config Providers
-pub mod config;
-
 /// Authz Utils
 pub mod authz;
