+++ nakago_sea_orm/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nakago-sea-orm"
-version = "0.6.0"
+version = "0.8.0"
 authors = ["Brandon Konkle <brandon@konkle.us>"]
 edition = "2021"
 description = "A SeaORM integration for Nakago"
@@ -10,10 +10,20 @@ license = "MIT"
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
+axum = { version = "0.6", features = ["macros"] }
+figment = { version = "0.10", features = ["env"] }
 log = "0.4"
-nakago = "0.6"
+nakago = "0.7"
 pretty_env_logger = "0.5"
 rand = "0.8"
+sea-orm = { version = "0.12", features = [
+    "macros",
+    "mock",
+    "runtime-tokio-rustls",
+    "sqlx-postgres",
+    "with-chrono",
+    "with-json",
+], default-features = false }
 serde = { version = "1.0", features = ["derive"] }
 serde_derive = "1.0"
 serde_json = "1.0"
